chore: chart -> core
