+++ src/core/chart.ts
@@ -20,6 +20,7 @@ import {
   handleZoomstart,
   measureXAxis,
   measureYAxis,
+  recalculateScale,
 } from "./helpers/chart-event-helpers";
 import {
   plotAreaInteraction,
@@ -85,6 +86,7 @@ export const chart = (
     "redraw",
     "rightclick"
   );
+
   let isPinned = true;
   let isFreePan = false;
 
@@ -572,39 +574,3 @@ export const chart = (
 
   return chart;
 };
-
-function recalculateScale(
-  xTransform: () => ZoomTransform,
-  xScale: ScaleTime,
-  yScales: Record<string, ScaleLinear>,
-  id: string,
-  plotAreas: Record<string, PlotAreaInterface>,
-  plotAreaElements: any,
-  yZooms: any
-) {
-  const xr = xTransform().rescaleX(xScale);
-  const bounds = xr.domain() as [Date, Date];
-  const originalExtent = yScales[id].range();
-  const newExtent = plotAreas[id].extent(bounds);
-  const originalHeight = Math.abs(originalExtent[0] - originalExtent[1]);
-
-  const newHeight = Math.abs(
-    yScales[id](newExtent[1]) - yScales[id](newExtent[0])
-  );
-
-  plotAreaElements[id].call(
-    yZooms[id].transform,
-    zoomIdentity
-      .translate(0, originalHeight / 2)
-      .scale(originalHeight / (1.3 * newHeight))
-      .translate(
-        0,
-        -(
-          yScales[id](newExtent[0]) -
-          0.1 * newHeight +
-          yScales[id](newExtent[1]) +
-          0.2 * newHeight
-        ) / 2
-      )
-  );
-}
