+++ src/core/plot-area-interaction.ts
