+++ src/core/plot-area.ts
