+++ src/core/x-axis-interaction.ts
